+++ src/test/java/lotto/domain/PurchasedLottoTest.java
@@ -1,8 +1,6 @@
 package lotto.domain;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.Arrays;
 import java.util.List;
